Removed the bower_components folder from the repository
The bower_components folder is filled with the dependencies necessary
for our project, but they change on a regular basis. It is best to
list the necessary dependencies inside the bower.json file and allow
the end user of the repository to pull all the up to date dependencies
with a simple "bower install" on the command line.